Programming/back-end/pyproject.toml
[tool.poetry]
name = "programming-question-generator-back-end"
version = "0.1.0"
description = "Back end for the project programming question generator"
authors = ["SajidAhmed0 <mjsahmed205@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
python-dotenv = "^1.0.1"
langchain = "^0.3.9"
langchain-community = "^0.3.9"
langchain-groq = "^0.2.1"
djangorestframework = "^3.15.2"
docker = "^7.1.0"
pyexecjs = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
